requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
opencv-python>=4.8.0
albumentations>=1.3.0
numpy<2
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
grad-cam>=1.4.0
PyYAML>=6.0
tqdm>=4.65.0
